refactor: move createPostfix into utils
