+++ lib/index.js
@@ -2,7 +2,7 @@ const WebSocket = require('ws')
 const EventEmitter = require('events')
 const request = require('request')
 const crypto = require('crypto')
-const { isObject, sleep } = require('./utils')
+const { createPostfix, isObject, sleep } = require('./utils')
 
 let api = function Bitvavo() {
   let base = 'https://api.bitvavo.com/v2'
@@ -56,21 +56,6 @@ let api = function Bitvavo() {
     }
     let signature = crypto.createHmac('sha256', apiSecret).update(string).digest('hex')
     return signature
-  }
-
-  const createPostfix = function (options) {
-    let params = Object.keys(options)
-      .reduce((a, k) => {
-        if (options[k].length !== 0) {
-          a.push(k + '=' + encodeURIComponent(options[k]))
-        }
-        return a
-      }, [])
-      .join('&')
-    if (Object.keys(options).length > 0) {
-      params = '?' + params
-    }
-    return params
   }
 
   const waitForSocketPrivate = function (socket, callback) {
