+++ lib/utils.js
@@ -1,4 +1,23 @@
 /**
+ * @param {{ [key: string]: string | number | boolean }} options
+ * @returns {string}
+ */
+exports.createPostfix = function createPostfix(options) {
+  let params = Object.keys(options)
+    .reduce((a, k) => {
+      if (options[k].length !== 0) {
+        a.push(k + '=' + encodeURIComponent(options[k]))
+      }
+      return a
+    }, [])
+    .join('&')
+  if (Object.keys(options).length > 0) {
+    params = '?' + params
+  }
+  return params
+}
+
+/**
  * @param {unknown} parameter
  * @returns {boolean}
  */
